Test PUT Layer

+++ src/layman/geoserver/wms.py
@@ -19,7 +19,8 @@ def update_layer(username, layername, layerinfo):
 
 
 def delete_layer(username, layername):
-    try:
+    info = get_layer_info(username, layername)
+    if info:
         r = requests.delete(
             urljoin(LAYMAN_GS_REST_WORKSPACES, username +
                     '/layers/' + layername),
@@ -32,9 +33,6 @@ def delete_layer(username, layername):
         # app.logger.info(r.text)
         r.raise_for_status()
         g.pop(FLASK_WMS_PROXY_KEY, None)
-    except Exception:
-        traceback.print_exc()
-        pass
     return {}
